Update apm.py
Add option to not print output from APM cmd. In a perfect world, I think it might be best to adjust the return value of method to return a tuple of `(response, output)` so the caller can do what they will with the output, or something similar, but that will likely cause a fair amount of refactoring.

This change should keep the default behavior while allowing to opt out from printing input, for example when running APM from Gekko. I haven't spent an inordinate amount of time thinking over this change, so feedback is certainly welcome :)